Restructure for consistency
