+++ domains/authorization/Authorization.API/API.IntegrationTests/API.IntegrationTests/Setup/IntegrationTestFixture.cs
@@ -1,4 +1,4 @@
-using API.Configuration;
+using API.Options;
 using Testcontainers.RabbitMq;
 
 namespace API.IntegrationTests.Setup;
