+++ domains/authorization/Authorization.API/API.IntegrationTests/API.IntegrationTests/Setup/TestWebApplicationFactory.cs
@@ -4,8 +4,8 @@ using System.Security.Claims;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
 using API.Authorization.Controllers;
-using API.Configuration;
 using API.Models;
+using API.Options;
 using FluentAssertions;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Hosting;
